la clef s'ajoute 1 seule fois dans l'inventaire

+++ game.js
@@ -1234,16 +1234,12 @@ class Board {
     }
 
     deleteKey(abs, ord, game) {
-        
         this.rows[abs][ord].image = null;
         this.rows[abs][ord].imgsrc = null;
         this.rows[abs][ord].walkable = true;
         let currentPlayer = game.currentPlayer
         currentPlayer.items.push(new Clef())
         updateInventoryTest(game);
-        
-        
-
     }
 
 
@@ -1932,7 +1928,7 @@ class Game {
 
     checkPos() {
 
-        if (this.currentPlayer.position[0] == 36 && this.currentPlayer.position[1] == 8) {
+        if (this.currentPlayer.position[0] == 36 && this.currentPlayer.position[1] == 8 && this.currentPlayer.hasKey == false) {
             this.board.deleteKey(36, 8, this);
             this.currentPlayer.hasKey = true;
         }
@@ -2126,7 +2122,7 @@ function updateUI(game) {
 
 //     console.log(inventory);
 }
-
+let compteurClef = 0;
 let game;
 document.getElementById('start-game').addEventListener('click', () => {
     game = new Game([new Player("Joueur 1"), new Player("Joueur 2")], new Board(10), new Inventory(10));
